feat: Add Django admin configurations for Instagram models and introduce Antigravity development workflows.
