+++ instagram/admin/story.py
@@ -1,0 +1,57 @@
+from django.contrib import admin
+from unfold.admin import ModelAdmin
+
+from instagram.models import Story
+
+
+@admin.register(Story)
+class StoryAdmin(ModelAdmin):
+    list_display = [
+        "story_id",
+        "user",
+        "created_at",
+        "story_created_at",
+    ]
+    list_filter = [
+        "user",
+        "created_at",
+        "story_created_at",
+    ]
+    search_fields = [
+        "story_id",
+        "user__username",
+    ]
+    readonly_fields = [
+        "story_id",
+        "user",
+        "created_at",
+        "story_created_at",
+        "raw_api_data",
+        "blur_data_url",
+    ]
+    fieldsets = (
+        (
+            "General",
+            {
+                "fields": (
+                    ("story_id", "user"),
+                    "story_created_at",
+                    "thumbnail",
+                    "media",
+                    "blur_data_url",
+                ),
+                "classes": ["tab"],
+            },
+        ),
+        (
+            "Metadata",
+            {
+                "fields": (
+                    "created_at",
+                    "raw_api_data",
+                ),
+                "classes": ["tab"],
+            },
+        ),
+    )
+    ordering = ["-created_at"]
