+++ instagram/admin/user.py
@@ -8,11 +8,7 @@ from simple_history.admin import SimpleHistoryAdmin
 from unfold.admin import ModelAdmin
 from unfold.decorators import action
 
-from .models import Story
-from .models import StoryCredit
-from .models import StoryCreditPayment
-from .models import User
-from .models import UserUpdateStoryLog
+from instagram.models import User
 
 
 @admin.register(User)
@@ -129,151 +125,3 @@ class InstagramUserAdmin(SimpleHistoryAdmin, ModelAdmin):
             )
 
         return redirect(reverse("admin:instagram_user_change", args=(object_id,)))
-
-
-@admin.register(UserUpdateStoryLog)
-class UserUpdateStoryLogAdmin(ModelAdmin):
-    list_display = [
-        "user",
-        "status",
-        "message",
-        "created_at",
-        "updated_at",
-    ]
-    list_filter = [
-        "status",
-        "created_at",
-        "updated_at",
-    ]
-    search_fields = ["user__username"]
-    readonly_fields = [
-        "user",
-        "status",
-        "message",
-        "created_at",
-        "updated_at",
-    ]
-    ordering = ["-created_at"]
-
-
-@admin.register(Story)
-class StoryAdmin(ModelAdmin):
-    list_display = [
-        "story_id",
-        "user",
-        "created_at",
-        "story_created_at",
-    ]
-    list_filter = [
-        "user",
-        "created_at",
-        "story_created_at",
-    ]
-    search_fields = [
-        "story_id",
-        "user__username",
-    ]
-    readonly_fields = [
-        "story_id",
-        "user",
-        "created_at",
-        "story_created_at",
-        "raw_api_data",
-    ]
-    fieldsets = (
-        (
-            "General",
-            {
-                "fields": (
-                    ("story_id", "user"),
-                    "story_created_at",
-                    "thumbnail",
-                    "media",
-                ),
-                "classes": ["tab"],
-            },
-        ),
-        (
-            "Metadata",
-            {
-                "fields": (
-                    "created_at",
-                    "raw_api_data",
-                ),
-                "classes": ["tab"],
-            },
-        ),
-    )
-    ordering = ["-created_at"]
-
-
-@admin.register(StoryCredit)
-class StoryCreditAdmin(SimpleHistoryAdmin, ModelAdmin):
-    list_display = [
-        "user",
-        "credit",
-        "created_at",
-        "updated_at",
-    ]
-    list_filter = [
-        "created_at",
-        "updated_at",
-    ]
-    search_fields = ["user__username"]
-    readonly_fields = [
-        "user",
-        "credit",
-        "created_at",
-        "updated_at",
-    ]
-    fieldsets = (
-        (
-            "General",
-            {
-                "fields": (
-                    ("user", "credit"),
-                    ("created_at", "updated_at"),
-                ),
-            },
-        ),
-    )
-    ordering = ["-created_at"]
-
-
-@admin.register(StoryCreditPayment)
-class StoryCreditPaymentAdmin(ModelAdmin):
-    list_display = [
-        "story_credit",
-        "payment",
-        "credit",
-        "created_at",
-        "updated_at",
-    ]
-    list_filter = [
-        "created_at",
-        "updated_at",
-    ]
-    search_fields = [
-        "story_credit__user__username",
-        "payment__reference",
-    ]
-    readonly_fields = [
-        "story_credit",
-        "payment",
-        "credit",
-        "created_at",
-        "updated_at",
-    ]
-    fieldsets = (
-        (
-            "General",
-            {
-                "fields": (
-                    ("story_credit", "payment"),
-                    "credit",
-                    ("created_at", "updated_at"),
-                ),
-            },
-        ),
-    )
-    ordering = ["-created_at"]
